requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aiosignal==1.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
future==0.18.2
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
huggingface-hub==0.4.0
idna==3.3
idna-ssl==1.1.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
Markdown==3.3.7
MarkupSafe==2.0.1
mistune==0.8.4
multidict==5.2.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauthlib==3.2.0
packaging==21.3
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.19.4
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyDeprecate==0.3.1
Pygments==2.12.0
pyparsing==3.0.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytorch-lightning==1.5.10
PyYAML==6.0
pyzmq==23.2.0
regex==2022.7.9
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.53
Send2Trash==1.8.0
six==1.16.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.12.1
testpath==0.6.0
tokenizers==0.12.1
torch==1.10.1+cu111
torchaudio==0.10.1+rocm4.1
torchmetrics==0.8.2
torchvision==0.11.2+cu111
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0
